backend/requirements.txt
fastapi==0.119.1
uvicorn==0.38.0
pydantic-ai==1.3.0
python-dotenv==1.1.1
logfire==4.14.1
httpx==0.28.1
